* homedir * catching exceptions by reference * fixing print outs * catching exceptions when property values cannot be set when loading from settings files

+++ Application/src/tracker/misc/MemoryStats.cpp
@@ -348,12 +348,12 @@ void MemoryStats::print() const {
         sorted.insert(key);
     std::vector<std::tuple<std::string, FileSize>> vec;
     for(auto &key : sorted)
-        vec.push_back({key, FileSize{sizes.at(key)}});
+        vec.push_back({key, FileSize{size_t(sizes.at(key))}});
     
     auto str = prettify_array(Meta::toStr(vec));
     auto id_str = id == Idx_t(std::numeric_limits<uint32_t>::max()-1) ? std::string("overall") : (!id.valid() ? "<empty>" : Meta::toStr(id));
     
-    cmn::print(id_str.c_str(), ": ", FileSize{bytes},"\n", str.c_str());
+    cmn::print(id_str.c_str(), ": ", FileSize{size_t(bytes)},"\n", str.c_str());
 }
 
 }
